// src/app/_services/swagger-auto-generated/model/spaceOverviewDto.ts
/**
 * HammerOn API: Easy control the band's songs
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 0.0.5
 * Contact: kb.developer.br@gmail.com
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { EventOverviewDto } from './eventOverviewDto';
import { MusicOverviewDto } from './musicOverviewDto';
import { UserOverviewDto } from './userOverviewDto';

export interface SpaceOverviewDto { 
    createdBy?: string;
    eventList?: Array<EventOverviewDto>;
    musicList?: Array<MusicOverviewDto>;
    spaceId?: string;
    spaceName?: string;
    userList?: Array<UserOverviewDto>;
}